Create the Access infrastructure.
PPP has four kinds of archived entities:
Forums (config), Topic, Post, Member (content).

Every archived entities can be classified, requiring
a special permission to view them.

This access is hierarchic. Viewing a post thus requires access to its
author and the topic it was posted in, which in turn requires access
to *its* author and the forum it was posted in.